draft of crowdfunding contract + traces
